changes that affect MASTER
merge this changes to Master. New processExp that allows modular running of different algorithms

# Conflicts:
#	.DS_Store
#	FociAnalysis/.DS_Store
#	Morphometrics_v2/.DS_Store
#	processExp.m

+++ processExp.m
@@ -22,36 +22,46 @@
 function processExp()
 % Steps in the pipeline: 
 
-% 1. Preparing the files
-% 2. Segmentation
-% 3. Morphometrics analysis
-% 4. Foci detection and Cell Cycle analysis
+%%  MODIFY PIPELINE ACCORDING TO EXPERIMENT SETTINGS AND SAVE IT IN EXPERIMENT FOLDER
+
+function processExp(preprocess,naming,supersegger,cleanup,imag2stack,morpho,fociCalc)
+
+% Steps of the pipeline: 
+%Boolean variables to decide which steps of the pipeline to run:
+
+% 1. Preparing files           - preprocess
+% 2. Convert file names        - naming
+% 3. SuperSegger Segmentation  - supersegger
+% 4. Clean up  files           - cleanup
+% 5. Convert images to stack   - imag2stack
+% 6. Morphometrics             - morpho
+% 7. Foci analysis             - fociCalc
 
 tic
 %% 1. Preparing tif files using MIJ (Fiji-Matlab interface)
 
-% Select folder with ND2 files
+if preprocess  
+    
+    % Select folder with ND2 files
+    disp('Select Folder with ND2 files:');
+    dirname = uigetdir();
+    dirname = fixDir(dirname);
 
-disp('Select Folder with ND2 files');
-dirname = uigetdir();
-dirname = fixDir(dirname);
+    %The macro converts ND2 files to tif for two separate channels. It performs 
+    %gaussian blur on the green channel if desired (check the macro code)
 
+    macroFile1='ConvertND2toTif.txt';
 
-%The macro converts ND2 files to tif for two separate channels. It performs 
-%gaussian blur on the green channel
+    % Call MIJ for preprocessing:
+    filepathMacro = getMacroPath(); %Macro Files path
 
-macroFile1='ConvertND2toTif.txt';
+    % Select frames to analyze
+    t_start=1;
+    t_end=89;
 
-% Call MIJ for preprocessing:
-filepathMacro = getMacroPath(); %Macro Files path
-
-% Select frames to analyze
-t_start=1;
-t_end=89;
-
-%calls MIJ to run the Fiji macro with arguments
-args=strcat(dirname,';',num2str(t_start),';',num2str(t_end)); %group arguments
-runMacro([filepathMacro,macroFile1],args); 
+    %calls MIJ to run the Fiji macro with arguments
+    args=strcat(dirname,';',num2str(t_start),';',num2str(t_end)); %group arguments
+    runMacro([filepathMacro,macroFile1],args); 
 
 %% Converting to SuperSegger naming format
 %
@@ -67,20 +77,30 @@ runMacro([filepathMacro,macroFile1],args);
 % t00002xy001c1.tif
 % etc ... without basename
 
-dirname =  ([dirname,'Analysis',filesep]); %set analysis folder
-cd(dirname);
-%Set conversion parameters based on your file names:
-basename = '';
-timeFilterBefore ='_t';
-timeFilterAfter = '.t' ;
-xyFilterBefore='';
-xyFilterAfter='xy';
-channelNames = {'C=0','C=1'};
+if naming  
+    
+    if ~exist( 'dirname','var')
+        dirname= pwd;
+        dirname = fixDir(dirname);
+        dirname =  ([dirname,'Analysis',filesep]);
+    end        
+    
+    dirname =  ([dirname,'Analysis',filesep]);
+    cd(dirname); % **This is important to keep**
 
-convertImageNames(dirname, basename, timeFilterBefore, ...
-    timeFilterAfter, xyFilterBefore,xyFilterAfter, channelNames )
+    %Set conversion parameters based on your file names:
+    basename = '';
+    timeFilterBefore ='_t';
+    timeFilterAfter = '.t' ;
+    xyFilterBefore='';
+    xyFilterAfter='xy';
+    channelNames = {'C=0','C=1'};
 
-%% Setting the segmentations constants for your bacteria and micrscope resolution
+    convertImageNames(dirname, basename, timeFilterBefore, ...
+        timeFilterAfter, xyFilterBefore,xyFilterAfter, channelNames )
+
+end
+%% 3. Setting segmentation constants and run SuperSegger
 %
 % Using correct resolution ensures correct pixel size and segmentation
 % constants if you do not know which constants to use you can run
@@ -94,18 +114,24 @@ convertImageNames(dirname, basename, timeFilterBefore, ...
 %[~, list] = getConstantsList;
 % list'
 
-res = '60XBsS750fil2'; %These constants have been optimized for B.subtilis filaments
+if supersegger    
+    if ~exist( 'dirname','var')
+        dirname= pwd;
+        dirname = fixDir(dirname);
+        dirname =  ([dirname,'Analysis',filesep]);
+    end        
 
 %% Paralell Processing Mode
 
 % to run code in parallel mode must have the parallel processing toolbox,
 % for convenience default is false (non-parallel)
 
-parallel_flag = true;
+    parallel_flag = true;
 
 %% Loading SuperSegger Constants
 
-CONST = loadConstants(res,parallel_flag) ;
+    res = '60XBsS750fil2'; %These constants have been optimized for B.subtilis filaments
+    CONST = loadConstants(res,parallel_flag) ;
 
 %% Modifying specific SuperSegger constant options
 
@@ -118,45 +144,45 @@ CONST = loadConstants(res,parallel_flag) ;
 % Alignment constants for multiple channel files: (Used bead images to
 % perform this registration)
 
-% Constants Calarco Microscope:
-CONST.imAlign.GFP     = [ 0.0000    0.0000    0.0000    0.0000];
-CONST.imAlign.mCherry = [0.04351    -0.0000    0.7200    0.5700]; 
-% *last two values are X and Y shifts.
+    % Constants Calarco Microscope:
+    CONST.imAlign.Phase     = [ 0.0000    0.0000    0.0000    0.0000];
+    CONST.imAlign.GFP       = [ 0.0000    0.0000    0.0000    0.0000];
+    CONST.imAlign.mCherry   = [0.04351   -0.0000    0.7200    0.5700]; 
+    CONST.imAlign.DAPI      = [0.0000     0.0000    0.0000    0.0000]; 
 
-% CONST.imAlign.DAPI    = [-0.0354   -0.0000    1.5500   -0.3900]; these
-% are super segger values
-% 
+% *last two values are X and Y shifts.
+ 
 % HOW TO OBTAIN REGISTRATION VALUES:
 % % im_base = imread('path/of/brightfield/in/non/shifted/channel.tif')
 % % im_shifted = imread('path/of/brightfield/in/shifted/channel.tif')
 % % [out,~,~] = intAlignIm(im_shifted, im_base, 100 ); %The values in 'out' are the four values needed for the shifted channel.
 % 
 % Channel order for registration:
-CONST.imAlign.out = {CONST.imAlign.GFP,...     % c1 channel (this case is Phase or Bright field)
-                    CONST.imAlign.GFP,...      % c2 channel name
-                    CONST.imAlign.mCherry,...  % c3 channel name
-                    CONST.imAlign.GFP};        % c4 channel name
-                
+    CONST.imAlign.out = {CONST.imAlign.Phase,...     % c1 channel (this case is Phase or Bright field)
+                        CONST.imAlign.GFP,...      % c2 channel name
+                        CONST.imAlign.mCherry,...  % c3 channel name
+                        CONST.imAlign.DAPI};        % c4 channel name
+
 %% Foci detection settings (For use only in SuperSegger)
 
-CONST.trackLoci.numSpots = [5 0]; % Max number of foci to fit in each fluorescence channel (default = [0 0])
-CONST.trackLoci.fluorFlag = false ;    % compute integrated fluorescence (default = true)
-CONST.trackOpti.NEIGHBOR_FLAG = false; % calculate number of neighbors (default = false)
-CONST.imAlign.AlignChannel = 1; % change this if you want the images to be aligned to fluorescence channel
-CONST.view.fluorColor = {'b','g','r'}; %Set the color for plotting different channels (in order)
+    CONST.trackLoci.numSpots = [5 0]; % Max number of foci to fit in each fluorescence channel (default = [0 0])
+    CONST.trackLoci.fluorFlag = false ;    % compute integrated fluorescence (default = true)
+    CONST.trackOpti.NEIGHBOR_FLAG = false; % calculate number of neighbors (default = false)
+    CONST.imAlign.AlignChannel = 1; % change this if you want the images to be aligned to fluorescence channel
+    CONST.view.fluorColor = {'b','g','r'}; %Set the color for plotting different channels (in order)
 
 %% Segmentation Filtering options in SuperSegger (modify if you see problems in detection)
 
-%The first two options (PEBBLE_CONST, INTENSITY_DIF) are for the remove_debris variable. 
-CONST.superSeggerOpti.PEBBLE_CONST = 1.3; %Default 1.5 for 60XEcM9
-CONST.superSeggerOpti.INTENSITY_DIF = 0.3; %Default 0.15 for 60XEcM9
-CONST.superSeggerOpti.remove_microcolonies =false; %Default is 1. It prevents deleting clusters of cells.
-CONST.superSeggerOpti.remove_debris = 1; %Turn off it is deleting cells
-CONST.superSeggerOpti.MAX_WIDTH = 1e15; %Set this high to prevent filaments to be split 
-CONST.seg.OPTI_FLAG = false; %To avoid segmenting cells by shape
+    %The first two options (PEBBLE_CONST, INTENSITY_DIF) are for the remove_debris variable. 
+    CONST.superSeggerOpti.PEBBLE_CONST = 1.3; %Default 1.5 for 60XEcM9
+    CONST.superSeggerOpti.INTENSITY_DIF = 0.3; %Default 0.15 for 60XEcM9
+    CONST.superSeggerOpti.remove_microcolonies =false; %Default is 1. It prevents deleting clusters of cells.
+    CONST.superSeggerOpti.remove_debris = 1; %Turn off it is deleting cells
+    CONST.superSeggerOpti.MAX_WIDTH = 1e15; %Set this high to prevent filaments to be split 
+    CONST.seg.OPTI_FLAG = false; %To avoid segmenting cells by shape
 
-% this helps cleanup death cells, fragments etc..
-CONST.trackOpti.REMOVE_STRAY = true;
+    % this helps cleanup death cells, fragments etc..
+    CONST.trackOpti.REMOVE_STRAY = true;
 
 % Other constants can be modified using the trainingGui and the sliders,
 % specially threshold1 and threshold2 and blur options can improve
@@ -165,16 +191,16 @@ CONST.trackOpti.REMOVE_STRAY = true;
 
 %% Time step and Pixel size settings:
 
-CONST.getLocusTracks.PixelSize =0.1; %Pixel size in Andor CCD 6.45 um divided by 60X
-CONST.getLocusTracks.TimeStep = 5; %time between frames, used for analysis
+    CONST.getLocusTracks.PixelSize =0.1; %Pixel size in Andor CCD 6.45 um divided by 60X
+    CONST.getLocusTracks.TimeStep = 5; %time between frames, used for analysis
 
 %% Options to remove CellAsic pillars:
 
-% CONST.superSeggerOpti.remove_pillars.flag = true;
-% CONST.superSeggerOpti.remove_pillars.radius = 2;
-% CONST.superSeggerOpti.remove_pillars.cut = 0.05;
-% CONST.superSeggerOpti.remove_pillars.Area_Cut = 700;
-% CONST.superSeggerOpti.remove_pillars.debug = false;
+    % CONST.superSeggerOpti.remove_pillars.flag = true;
+    % CONST.superSeggerOpti.remove_pillars.radius = 2;
+    % CONST.superSeggerOpti.remove_pillars.cut = 0.05;
+    % CONST.superSeggerOpti.remove_pillars.Area_Cut = 700;
+    % CONST.superSeggerOpti.remove_pillars.debug = false;
 
 %% Skip Frames for Segmentation
 
@@ -182,15 +208,15 @@ CONST.getLocusTracks.TimeStep = 5; %time between frames, used for analysis
 % during segmentation to increase processing speed. Fluorescence images 
 % will not be skipped. 
 
-skip = 1;  % segment every frame
-%skip = 5; % segment every fifth phase image
+    skip = 1;  % segment every frame
+    %skip = 5; % segment every fifth phase image
 
 %% Clean previous segmented data
 
 % If set to true, will begin processing on aligned images; if false, will 
 % try to restart processing at last successful function (default = false)
 
-cleanflag = 1;
+    cleanflag = 1;
 
 %% 2. Running segmentation in SuperSegger
 
@@ -207,29 +233,48 @@ cleanflag = 1;
 % 9 : clist
 % 10 : cell files
 
-% in MorphoSegger, only Alignment, Segmentation and Stripping are used and
-% indicated in the variable startEnd.
+    % in MorphoSegger, only Alignment, Segmentation and Stripping are used and
+    % indicated in the variable startEnd.
 
-startEnd = [1 3];
-BatchSuperSeggerOpti( dirname, skip, cleanflag, CONST,startEnd);
+    startEnd = [1 3];
+    BatchSuperSeggerOpti( dirname, skip, cleanflag, CONST,startEnd);
 
 %clearvars -except dirname filepathMacro
 
 %% Clean up files (raw_im *.tif,*.mat)
 
-%Delete original and raw_im folders:
-disp('Deleting files...')
-rmdir( [dirname,filesep,'original',filesep],'s' ); 
-rmdir( [dirname,filesep,'raw_im',filesep],'s' );
 
+end
+%% 4. Clean up files (raw_im *.tif, original, *.tif)
 
-%% Converting images to stack 
+if cleanup
+     if ~exist( 'dirname','var')
+        dirname= pwd;
+        dirname = fixDir(dirname);
+        dirname =  ([dirname,'Analysis',filesep]);
+    end     
+    
+    %Delete original and raw_im folders:
+    disp('Deleting files...')
+    rmdir( [dirname,filesep,'original',filesep],'s' ); 
+    rmdir( [dirname,filesep,'raw_im',filesep],'s' );
 
-disp('Converting Images to Stack');
-macroFile2='tifImagesToStack.txt';
-runMacro([filepathMacro,macroFile2],dirname);
+end
+%% 5. Converting images to stack 
 
-%% 3. Run Morphometrics for Pill Mesh calculation 
+if imag2stack   
+     if ~exist( 'dirname','var')
+        dirname= pwd;
+        dirname = fixDir(dirname);
+        dirname =  ([dirname,'Analysis',filesep]);
+     end     
+    
+    disp('Converting Images to Stack...');
+    macroFile2='tifImagesToStack.txt';
+    runMacro([filepathMacro,macroFile2],dirname);
+
+end
+%% 6. Run Morphometrics for Pill Mesh calculation 
 
 % Morphometrics is a robust pipeline that interpolates cell contours at
 % subpixel resolution using PSICIC. It uses Oufti routines for segmentation
@@ -237,43 +282,60 @@ runMacro([filepathMacro,macroFile2],dirname);
 % Morphometrics to quickly re-segment the binary masks, contour
 % fitting, mesh calculation together with basic lineage tracking.
 
-paramName ='Morphometrics_prefs_mask_CL'; %Select parameter file 
-params = loadParams( paramName );
-% 
-% %List of most frequently changed parameters, modify here for different
-% %types of images
-% params.v_imtype = 2;        % 1 = Phase; 2 = Fluorescence (internal); 3 = Fluorescence (peripheral)      
-% params.v_method = 3;        % 1 = Gradient Segmentation; 2 = Laplacian Segmentation; 
-%                             % 3 = Adaptive Threshold Segmentation; 4 = Canny Segmentation
-% params.v_simplethres=1;     % Simple threshold 
-% params.f_areamin = 10;     % Min region size
-% params.f_areamax = 200000;  % Max region size
-% params.v_prox = 0;          % Cells are in proximity
-% params.v_exclude=0;         % Exclude edge objects
-% params.f_hmin_split=2;      % Cut distance (pxls)
-% params.v_save = 1;          % Save output
-% params.v_mt_mesh=1;         % Pill mesh
-% params.v_falsepos = 1;      % Reject false positives
-% params.f_int_rej = 3;       % False positive rejection parameter
-% % Tracking parameters:
-% params.f_pert_same = 0.55;  % Fractional overlap
-% params.f_frame_diff = 4;    % Frame overlap
-% %workers = 6;                % Number of workers for parallel job
-% 
-disp('Running Morphometrics in parallel')
-run_parallel(dirname,params);
+if morpho    
+     if ~exist( 'dirname','var')
+        dirname= pwd;
+        dirname = fixDir(dirname);
+        dirname =  ([dirname,'Analysis',filesep]);
+    end     
 
-%% 4. Foci calculation - Diego's pipeline
+    paramName ='Morphometrics_prefs_mask_CL'; %Select parameter file 
+    params = loadParams( paramName );
+    % 
+    %List of most frequently changed parameters, modify here for different
+    %types of images
+    params.v_imtype = 2;        % 1 = Phase; 2 = Fluorescence (internal); 3 = Fluorescence (peripheral)      
+    params.v_method = 3;        % 1 = Gradient Segmentation; 2 = Laplacian Segmentation; 
+                                % 3 = Adaptive Threshold Segmentation; 4 = Canny Segmentation
+    params.v_simplethres=1;     % Simple threshold 
+    params.f_areamin = 10;     % Min region size
+    params.f_areamax = 200000;  % Max region size
+    params.v_prox = 0;          % Cells are in proximity
+    params.v_exclude=0;         % Exclude edge objects
+    params.f_hmin_split=2;      % Cut distance (pxls)
+    params.v_save = 1;          % Save output
+    params.v_mt_mesh=1;         % Pill mesh
+    params.v_falsepos = 1;      % Reject false positives
+    params.f_int_rej = 3;       % False positive rejection parameter
+    % Tracking parameters:
+    params.f_pert_same = 0.55;  % Fractional overlap
+    params.f_frame_diff = 4;    % Frame overlap
+    %workers = 6;                % Number of workers for parallel job
 
-paramFit=50;     % Consecutive points for a single cell to include the trajectory
-Dparameter=65;   % Threshold to detect if it is foci in Diego's algorithm
-exp_cut=65;      % Takes only 65 pixels onwards to fit the exponential. Those points are not included in the gc_fit plot but the fit itself starts on point 65
-noiseTh=8;       % Noise threshold for wavelet detection
+    disp('Running Morphometrics in parallel...')
+    run_parallel(dirname,params);
 
-disp('Running Foci Analysis')
-run_fociAnalysis(dirname,paramFit,CONST.getLocusTracks.TimeStep,Dparameter,exp_cut,noiseTh)
+end
 
+%% 7. Foci calculation - Diego's pipeline
 
+if fociCalc
+     if ~exist( 'dirname','var')
+        dirname= pwd;
+        dirname = fixDir(dirname);
+        dirname =  ([dirname,'Analysis',filesep]);
+     end
+     
+    %Parameters    
+    paramFit=50;     % Consecutive points for a single cell to include the trajectory
+    Dparameter=65;   % Threshold to detect if it is foci in Diego's algorithm
+    exp_cut=65;      % Takes only 65 pixels onwards to fit the exponential. Those points are not included in the gc_fit plot but the fit itself starts on point 65
+    noiseTh=8;       % Noise threshold for wavelet detection
+
+    disp('Running Foci Analysis...')
+    run_fociAnalysis(dirname,paramFit,CONST.getLocusTracks.TimeStep,Dparameter,exp_cut,noiseTh)
+
+end
 %% Shutting down parallel pool
 
 disp('Closing parallel pool...')
@@ -284,21 +346,7 @@ delete(poolobj);
 
 t1=toc;
 disp(['Finished in ' num2str(round(10*t1/60)/10) ' minutes.']);
-cd(dirname);
+
 end
-%% Unused code (might be useful for some other things
 
 
-%disp ('Do you want to cleanup folders?, Y/N [Y]:')
-% try
-%     answer=input('','s');
-%     if lower(answer) ~='y'
-%         disp ('Exiting BatchSuperSegger. Reset clean flag and rerun');
-%         return
-%     end
-% catch
-%     % can not use input  - in eval mode
-% end
-
-%copyfile( [dirname,filesep,'xy1',filesep,'seg',filesep,'*.tif'], [dirname,filesep,'morphometrics'] )
-% delete ([dirname_cell,'cell*.mat']);
